// SIGE/src/app/usuarios/interfaces/usuario.interfaces.ts
// Generated by https://quicktype.io

export interface Usuarios {
  msj:     string;
  data:    DatumUsuarios[];
  errores: any[];
}

export interface DatumUsuarios {
  id:         number;
  usuario:    string;
  contrasena: string;
  correo:     string;
  codigo:     string;
  fallido:    number;
  estado:     estadoList;
}

export enum estadoList{

  ACT='ACT',
  INA='INA',
  BLO= 'BLO'


}