// SIGE/src/app/usuarios/interfaces/usuId.interface.ts
// Generated by https://quicktype.io

export interface IDUsuarios {
  msj:     string;
  data:    DataIdUser;
  errores: any[];
}

export interface DataIdUser {
  id:         number;
  usuario:    string;
  contrasena: string;
  correo:     string;
  codigo:     string;
  fallido:    number;
  estado:     estadoList;
}
export enum estadoList{

  ACT='ACT',
  INA='INA',
  BLO= 'BLO'


}